+++ app/helpers/profiles_helper.rb
@@ -4,7 +4,7 @@
 module ProfilesHelper
   def profile_picture_tag(user, size: 128, style: nil)
     image_tag(
-      user.avatar_url(size),
+      profile_picture_member_path(user),
       alt: "",
       class: "rounded img-ignore-selection",
       size: "#{size}x#{size}",