+++ CHANGELOG.md
@@ -8,6 +8,8 @@
   - Changed APIs for subject events and event designs
 - **Design Changes**
   - Designs are now searchable by short name and slug on the design index
+- **Profile Changes**
+  - Users can now directly upload profile pictures
 - **Project Changes**
   - Added a task that checks calculated fields and provides a list of sheets
     that have a mismatch between the recomputed and stored value